fix dependencies and package structure
- move some dependencies to right place
- divide CoverChecker and console Launcher
