+++ cover-checker-console/pom.xml
@@ -32,5 +32,12 @@
             <artifactId>cover-checker-github</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+
+
+        <dependency>
+            <groupId>commons-cli</groupId>
+            <artifactId>commons-cli</artifactId>
+            <version>1.4</version>
+        </dependency>
     </dependencies>
 </project>
