+++ cover-checker-core/src/main/java/com/naver/nid/cover/util/Parameter.java
